requirements.txt
# Using Python 3.8.2
grpcio-tools==1.34.1
grpcio==1.34.1
lxml==4.6.2
numpy==1.19.4
protobuf==3.14.0

## deveelopment modules
ipdb